requirements.txt
pandas~=2.2.3
numpy~=1.26.4
faker~=37.1.0
sqlalchemy~=2.0.34
python-dotenv~=0.21.0
python-dotenv==0.21.0
mlflow==2.18.0